requirements.txt
opencv-python>=4.8.0
numpy>=1.24.0
streamlit>=1.28.0
pandas>=2.0.0
plotly>=5.17.0
pyyaml>=6.0
pillow>=10.0.0
scipy>=1.11.0
